Protokoll #272 Paket 4a: Mailtransport abgrenzen
